+++ imports/components/app/subscriptionManager/SubscriptionManagerConnect.test.js
@@ -13,12 +13,15 @@ describe('mapStateToProps Function', () => {
   it('substitutes an empty string if there is no team on the user', () => {
     const state = { app: {} };
     Meteor.loggedInUser = { profile: { team: 'a' } };
+    expect(mapStateToProps(state).loading).toBe(true);
     expect(mapStateToProps(state).subscriptions.teams[1]).toBe('a');
     expect(mapStateToProps(state).subscriptions.teamActivity[1]).toBe('a');
     expect(mapStateToProps(state).subscriptions.opportunityForecast[1]).toBe(
       'a'
     );
     Meteor.loggedInUser = null;
+    state.app.loading = false;
+    expect(mapStateToProps(state).loading).toBe(false);
     expect(mapStateToProps(state).subscriptions.teams[1]).toBe('');
     expect(mapStateToProps(state).subscriptions.teamActivity[1]).toBe('');
     expect(mapStateToProps(state).subscriptions.opportunityForecast[1]).toBe(
@@ -30,6 +33,6 @@ describe('mapDispatchToProps Function', () => {
   it('maps login dispatchers', () => {
     const dispatch = jest.fn();
     const props = mapDispatchToProps(dispatch);
-    expect(props).toEqual({});
+    expect(props.setLoading).not.toThrow();
   });
 });