+++ imports/components/pages/objectEditor/ObjectEditorConnect.js
@@ -4,7 +4,8 @@ import { Meteor } from 'meteor/meteor';
 import ObjectEditorContainer from './ObjectEditorContainer';
 import { setObjectEditorProp } from '../../../state/actions/objectEditorActionCreators';
 
-export const mapStateToProps = ({ objectEditor }) => ({
+export const mapStateToProps = ({ app, objectEditor }) => ({
+  loading: app.loading === false ? false : true,
   hasLoaded: objectEditor.hasLoaded || false,
   loadedValues: objectEditor.loadedValues || {},
 });